refactor: Clean up FGG3 migration follow-ups
Move timerStateLabel out of the API layer into a view-local constant,
drop dead sign-handling from UiTimestamp (timer/game durations are
never negative), and collapse the /login and /signup routes into a
single route with local tab state.
