+++ src/router/authGuard.ts
@@ -6,10 +6,7 @@ export const authGuard: NavigationGuard = async (to) => {
 	const authStore = useAuthStore()
 	const isLoggedIn = await authStore.getIsLoggedIn()
 
-	if (
-		typeof to.name === 'string' &&
-		[RouteName.Login, RouteName.Signup].includes(to.name as RouteName)
-	) {
+	if (to.name === RouteName.Login) {
 		return !isLoggedIn || { name: RouteName.Timer }
 	}
