+++ src/views/Timer/constants/timerStateLabel.ts
@@ -1,0 +1,8 @@
+import { TimerState } from '../../../api-facade/models/timers-models'
+
+export const timerStateLabel: Record<TimerState, string> = {
+	[TimerState.Created]: 'создан',
+	[TimerState.Running]: 'идёт',
+	[TimerState.Paused]: 'пауза',
+	[TimerState.Finished]: 'завершён',
+}
